initial release (#2)
* init push

* fix provider error

* specify full api group to avoid conflcts

* fixes outputs

+++ .github/scripts/validate-deploy.sh
@@ -33,6 +33,7 @@ BRANCH=$(jq -r '.branch // "main"' gitops-output.json)
 SERVER_NAME=$(jq -r '.server_name // "default"' gitops-output.json)
 LAYER=$(jq -r '.layer_dir // "2-services"' gitops-output.json)
 TYPE=$(jq -r '.type // "base"' gitops-output.json)
+CLUSTERID=$(jq -r '.clusterid // "mykafka"' gitops-output.json)
 
 mkdir -p .testrepo
 
@@ -48,7 +49,35 @@ validate_gitops_content "${NAMESPACE}" "${LAYER}" "${SERVER_NAME}" "${TYPE}" "${
 
 check_k8s_namespace "${NAMESPACE}"
 
-#check_k8s_resource "${NAMESPACE}" "deployment" "${COMPONENT_NAME}"
+# Operator check
+check_k8s_resource "${NAMESPACE}" "deployment" "strimzi-cluster-operator-v0.22.1"
+
+# cluster deploy check
+check_k8s_resource "${NAMESPACE}" "deployment" "maskafka-entity-operator"
+
+# wait for kafka server job to complete
+sleep 1m
+
+# check kafka cluster is in ready state
+kafkastatus=$(kubectl get kafkas.kafka.strimzi.io ${CLUSTERID} -n ${NAMESPACE} --no-headers -o custom-columns=":status.conditions[0].type")
+
+count=0
+until [[ "${kafkastatus}" = "Ready" ]] || [[ $count -eq 20 ]]; do
+  echo "Waiting for ${CLUSTERID} in ${NAMESPACE}"
+  count=$((count + 1))
+  kafkastatus=$(kubectl get kafkas.kafka.strimzi.io ${CLUSTERID} -n ${NAMESPACE} --no-headers -o custom-columns=":status.conditions[0].type")
+  sleep 60
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for ${CLUSTERID} to become ready in ${NAMESPACE}"
+  kubectl get all -n "${NAMESPACE}"
+  exit 1
+fi
+
+# wait for any cleanup to complete if needed
+sleep 1m
+
 
 cd ..
 rm -rf .testrepo
